# Improve list slice performance