I noticed that we could use `VECTOR_PTR_RO()` when slicing lists, rather than `VECTOR_ELT()`, resulting in a slight but sometimes noticeable speedup.

The diffs look bad, but its really just replacing `GET` with `CTYPE` and `CONST_DEREF`

I've stayed consistent with the `data` naming convention of the other slice macros for now. We can update them all later, but I felt it was better to be consistent at the moment.

```r
library(vctrs)

x <- rep(list(1), 1e8)
i <- seq_along(x) + 0L

bench::mark(vec_slice(x, i), iterations = 50)

# Master
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, i)    1.17s    1.24s     0.800     763MB    0.940

# This PR
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_slice(x, i)    998ms    1.03s     0.962     763MB     1.13
```